refactor(platform): /domains/ router refactor
