+++ platform/api/hexgate_api/domains/__init__.py
@@ -1,0 +1,1 @@
+"""Domain packages: one vertical slice per bounded context."""
