+++ platform/api/hexgate_api/domains/agents/__init__.py
@@ -1,0 +1,1 @@
+"""Agent registry, manifests, policy validation, and bundle reads."""
