+++ platform/api/hexgate_api/domains/audit/__init__.py
@@ -1,0 +1,1 @@
+"""Audit ingest + dashboard read endpoints."""
