+++ platform/api/hexgate_api/domains/chat/__init__.py
@@ -1,0 +1,1 @@
+"""WebSocket relay endpoints: SDK serve producer + dashboard chat consumer."""
