+++ platform/api/hexgate_api/domains/invitations/__init__.py
@@ -1,0 +1,1 @@
+"""Org invitations."""
