+++ platform/api/hexgate_api/domains/orgs/__init__.py
@@ -1,0 +1,1 @@
+"""Organization CRUD."""
