+++ platform/api/hexgate_api/domains/projects/__init__.py
@@ -1,0 +1,1 @@
+"""Project CRUD."""
