+++ platform/api/hexgate_api/domains/tokens/__init__.py
@@ -1,0 +1,1 @@
+"""Tokens & signing-key endpoints."""
