+++ platform/api/tests/test_tenant_isolation.py
@@ -21,7 +21,6 @@ from sqlalchemy.pool import StaticPool
 from sqlmodel import SQLModel
 from sqlmodel.ext.asyncio.session import AsyncSession
 
-from hexgate_api import main
 from hexgate_api.main import app
 from hexgate_api.models import Agent, Organization, OrganizationMember, Project, User
 from hexgate_api.services import (
@@ -158,7 +157,9 @@ async def client(session_factory) -> TestClient:
         async with session_factory() as session:
             yield session
 
-    app.dependency_overrides[main.get_session] = override_session
+    from hexgate_api.core.db import get_session
+
+    app.dependency_overrides[get_session] = override_session
     try:
         yield TestClient(app)
     finally:
